adjustments necessary to support Vanilla version

+++ wow-commons/src/main/java/wow/commons/repository/impl/parsers/stats/StatMatcher.java
@@ -63,7 +63,7 @@ public class StatMatcher {
 
 	public Attributes getParamStats() {
 		String value = evalParams(pattern.getParams().getAmount());
-		int amount = Integer.parseInt(value);
+		double amount = Double.parseDouble(value);
 		return pattern.getParams().getStatsSupplier().getAttributes(amount);
 	}
